LV Model add and stores mean in dict, replots when ROI is moved
